Allow to collect moving back task in open transition
Simulation:
 Task A starts(NEW_TASK) Task B, finishes
 Task B starts(NEW_TASK) Task C
   Task B calls moveTaskToBack in onPause
 Task C (declare as translucent and sleep 200ms in
         onCreate to delay transition)
   starts(NEW_TASK) Task B -> attempt to move task B to front
   finishes

Because the initial open transition may be done until the last
step, then moveTaskToBack for B will run after the last step.
Then expect B should be on top but the result is on home screen
because A and C are finished, and B is moved to back.

This change is kinda a workaround to mitigate the case of
deferred move-to-back. Because it is regular to have closing
targets in an open transition, the moving task can run with
the collecting transition directly.

Fix: 325645759
Test: atest TransitionTests#testDeferredMoveTaskToBack
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:8a5876d3cbab2b62e2294589d572480468063518 [formerly 7079ba9aad0012ae53cb689dc23740b99ed27d15])
Merged-In: I46d8e26110ebefea74e5dd06c85c003b12693416
Change-Id: I46d8e26110ebefea74e5dd06c85c003b12693416

Former-commit-id: 13eeb8e1cba0db3f286aa20d4d9222e399a1625a